--- src/main/java/Application.java
@@ -1,9 +1,0 @@
-import controller.BlackjackGame;
-
-public class Application {
-
-    public static void main(String[] args) {
-        BlackjackGame blackjackGame = new BlackjackGame();
-        blackjackGame.run();
-    }
-}
