+++ src/main/java/blackjack/domain/deck/Shape.java
@@ -1,0 +1,5 @@
+package blackjack.domain.deck;
+
+public enum Shape {
+    HEART, DIAMOND, CLOVER, SPADE;
+}
