+++ src/main/java/blackjack/dto/CardDto.java
@@ -1,0 +1,11 @@
+package blackjack.dto;
+
+import blackjack.domain.deck.Card;
+
+public record CardDto (String cardName){
+    public static CardDto from(Card card) {
+        String rankName = RankDto.from(card.getRank()).getRankName();
+        String shapeName = ShapeDto.from(card.getShape()).getShapeName();
+        return new CardDto(rankName + shapeName);
+    }
+}
