--- src/main/java/domain/deck/Shape.java
@@ -1,5 +1,0 @@
-package domain.deck;
-
-public enum Shape {
-    HEART, DIAMOND, CLOVER, SPADE;
-}
