--- src/main/java/dto/CardDto.java
@@ -1,11 +1,0 @@
-package dto;
-
-import domain.deck.Card;
-
-public record CardDto (String cardName){
-    public static CardDto from(Card card) {
-        String rankName = RankDto.from(card.getRank()).getRankName();
-        String shapeName = ShapeDto.from(card.getShape()).getShapeName();
-        return new CardDto(rankName + shapeName);
-    }
-}
