--- src/main/java/dto/DeckDto.java
@@ -1,14 +1,0 @@
-package dto;
-
-import domain.deck.Deck;
-import java.util.List;
-
-public record DeckDto (List<String> cardNames){
-    public static DeckDto from(Deck deck) {
-        List<String> cardNameList = deck.getCards().stream()
-                .map(CardDto::from)
-                .map(CardDto::cardName)
-                .toList();
-        return new DeckDto(cardNameList);
-    }
-}
