+++ src/test/java/blackjack/domain/RankTest.java
@@ -1,0 +1,21 @@
+package blackjack.domain;
+
+import blackjack.domain.deck.Rank;
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+class RankTest {
+
+    @Test
+    @DisplayName("에이스기 참임을 확인한다.")
+    void isAceSuccessTest() {
+        Assertions.assertThat(Rank.ACE.isAce()).isTrue();
+    }
+
+    @Test
+    @DisplayName("에이스기 거짓임을 확인한다.")
+    void isAceFailTest() {
+        Assertions.assertThat(Rank.SIX.isAce()).isFalse();
+    }
+}
